Branch three walks: the account first run made signs somebody in
The branch where a firm has no directory at all, end to end for the first
time. First run hashes the password and writes it; the dashboard makes the
account at start; somebody signs in with it, redeems the claim, and is this
deployment's first administrator. Then five wrong passwords and the right one
stops working.

This is the route that had two defects survive a year of green tests -- a
groups claim and an entire account, each collected by the wizard and read by
nobody. Both survived because nothing signed anybody in here. Something does
now.

It found a third within a minute of first running, which is the argument for
it. **The lockout was broken in Postgres and correct in memory.** The update
added two bare parameters, and `unknown + unknown` has no unique operator, so
every failed attempt errored at execution. The unit tests passed because they
run against the in-memory store; the branch that actually ships was refusing
sign-ins with a 503 and counting nothing. Cast now, with the reason in the
SQL.

Two more of the same family, both found by running it rather than reading it.
The blocking Postgres client makes a runtime of its own, so constructing it
inside the async one panics in a destructor -- a crash with no cause in the
log and a closed connection at the browser. It is built before the runtime
starts now, and authentication runs on the blocking pool, where Argon2 belongs
anyway: it is deliberately slow, and a handful of sign-ins would otherwise
stall every other request.

The hash the target uses is a fixed one, produced by this crate's own
`hash_password`. `e2e-first-run` asserts the Job writes a hash of that shape
and that the password reaches no Secret; this asserts the dashboard turns one
into a working account. The two halves meet at the format, and the Makefile
says so rather than leaving it to be discovered.

+++ crates/dashboard/src/accounts.rs
@@ -330,8 +330,13 @@ impl Accounts for InPostgres {
             .execute(
                 "UPDATE dashboard_local_account
                     SET failed_attempts = failed_attempts + 1,
+                        -- Cast, because two bare parameters added together
+                        -- give Postgres nothing to infer from: `unknown +
+                        -- unknown` has no unique operator, and the statement
+                        -- fails at execution rather than at compile time.
                         locked_until_ns = CASE
-                            WHEN failed_attempts + 1 >= $2 THEN $3 + $4
+                            WHEN failed_attempts + 1 >= $2::integer
+                            THEN $3::bigint + $4::bigint
                             ELSE locked_until_ns
                         END
                   WHERE name = $1",
